requirements.txt
ahk==0.9.0
certifi==2020.4.5.2
chardet==3.0.4
comtypes==1.1.7
cx-Freeze==6.1
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
MouseInfo==0.1.3
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
PyAutoGUI==0.9.50
PyDirectInput==1.0.1
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pynput==1.6.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.4
PyTweening==1.0.3
pywin32==228
pywinauto==0.6.8
requests==2.24.0
six==1.15.0
urllib3==1.25.9
